% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep_graph.R
\name{plot_deps_graph}
\alias{plot_deps_graph}
\title{Plot dependency graph}
\usage{
plot_deps_graph(
  pkg,
  dep_type = c("depends", "imports", "linkingto"),
  n_iter = 100,
  pak_res = NULL,
  gg = FALSE,
  lwd = 1,
  cex = 1,
  pad_h = 0.09,
  pad_w = 0.08,
  arw = 1,
  log_col_scale = FALSE
)
}
\arguments{
\item{pkg}{package string passed to \code{\link[pak:pkg_deps]{pak::pkg_deps}}}

\item{dep_type}{type(s) of dependencies to look up. Valid values are
\code{c("depends", "imports", "linkingto", "suggests")}}

\item{n_iter}{number of iterations for stress graph layout computation}

\item{pak_res}{a pre-computed result from
\code{\link[pak:pkg_deps]{pak::pkg_deps}}}

\item{gg}{If true, use ggplot2 + ggraph to draw the plot instead of base
graphics.}

\item{lwd}{line width}

\item{cex}{text size multiplication factor (see
\code{\link[graphics:par]{graphics::par}})}

\item{pad_h}{height padding}

\item{pad_w}{width padding}

\item{arw}{factor by which to lengthen/shorten arrowheads}

\item{log_col_scale}{if TRUE, set node colors to vary on a log scale with}
}
\value{
a ggplot
}
\description{
Plot a package's dependency graph, coloring each node by the
  number of packages it depends on.
}
\details{
If you include \code{"suggests"} among the dependency types to look
  up, be aware that suggests can be circular / cyclic. If this is detected,
  the node coloring will be turned off.

  Pre-computing the dependency lookup with
  \code{\link[pak:pkg_deps]{pak::pkg_dep}} and passing it to the
  \code{pak_res} argument can be handy when fiddling with graphical
  parameters. Also handy to avoid hitting the bundled GitHub PAT limits used
  by pak::pkg_deps().

  The arrows will be off if you resize the panel with the base graphics
  version. Either set \code{gg=TRUE} or set your desired graphics device
  size, then re-run your command.

  The layout in the base graphics version is stochastic - if it looks weird
  just run it again.
}
\examples{
# Using pkgcache in examples is not allowed, uncomment to run these interactively:
# plot_deps_graph("ggplot2") # ggplot2 has many downstream dependencies
# plot_deps_graph("rlang") # rlang has only one

}
